requirements.txt
fastapi
uvicorn
torch
torchvision 
ultralytics
opencv-python
numpy
pillow
python-multipart